sandbox: update sandbox
